# www/api/openapi.yaml
openapi: 3.0.3
info:
  title: RootWarden API
  description: |
    API REST du backend Flask de RootWarden — plateforme de gestion centralisee
    de serveurs Linux (cles SSH, mises a jour, iptables, scan CVE).

    **Authentification** : header `X-API-KEY` obligatoire sur toutes les routes (sauf /test).
    Le frontend PHP utilise `/api_proxy.php` qui injecte la cle automatiquement.
  version: 1.11.0
  contact:
    name: RootWarden
    url: https://github.com/Timikana/rootwarden
servers:
  - url: https://localhost:5000
    description: Backend direct (dev)
  - url: /api_proxy.php
    description: Via proxy PHP (production)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
...
security:
  - ApiKeyAuth: []
paths:
...
  /services/list:
    post:
      tags: [Services]
      summary: Liste des services systemd
      description: Liste tous les services systemd sur un serveur avec statut, categorie et enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id]
              properties:
                machine_id: { type: integer, example: 1 }
      responses:
        '200':
          description: Liste des services
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  services:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string, example: "nginx" }
                        active: { type: string, example: "active" }
                        sub: { type: string, example: "running" }
                        unit_file_state: { type: string, example: "enabled" }
                        description: { type: string }
                        category: { type: string, example: "web" }
                        protected: { type: boolean }
  /services/status:
    post:
      tags: [Services]
      summary: Detail d'un service
      description: Retourne le statut detaille (PID, memoire, uptime, description).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
      responses:
        '200':
          description: Detail du service
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  ActiveState: { type: string, example: "active" }
                  SubState: { type: string, example: "running" }
                  MainPID: { type: string, example: "1234" }
                  MemoryCurrent: { type: string, example: "905216" }
                  Description: { type: string }
  /services/start:
    post:
      tags: [Services]
      summary: Demarrer un service
      description: Demarre un service systemd. Bloque si service protege (sshd, dbus).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
      responses:
        '200':
          description: Resultat
        '403':
          description: Service protege
  /services/stop:
    post:
      tags: [Services]
      summary: Arreter un service
      description: Arrete un service systemd. Bloque si service protege.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
      responses:
        '200':
          description: Resultat
        '403':
          description: Service protege
  /services/restart:
    post:
      tags: [Services]
      summary: Redemarrer un service
      description: Redemarre un service systemd. Bloque si service protege.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
      responses:
        '200':
          description: Resultat
        '403':
          description: Service protege
  /services/enable:
    post:
      tags: [Services]
      summary: Activer un service au boot
      description: Active un service au demarrage (systemctl enable). Bloque si protege.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
      responses:
        '200':
          description: Resultat
        '403':
          description: Service protege
  /services/disable:
    post:
      tags: [Services]
      summary: Desactiver un service au boot
      description: Desactive un service au demarrage (systemctl disable). Bloque si protege.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
      responses:
        '200':
          description: Resultat
        '403':
          description: Service protege
  /services/logs:
    post:
      tags: [Services]
      summary: Logs journalctl d'un service
      description: Lit les dernieres lignes du journal systemd (journalctl -u).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, service]
              properties:
                machine_id: { type: integer, example: 1 }
                service: { type: string, example: "nginx" }
                lines: { type: integer, example: 100, default: 50 }
      responses:
        '200':
          description: Logs du service
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  logs: { type: string }
...